requirements.txt
Django
djangorestframework
mysqlclient
PyMySQL
pytz
